Add basic func test case
