+++ Cargo.toml
@@ -14,8 +14,8 @@ clap = { version = "4.2.7", features = ["derive"] }
 anyhow = { version = "1.0.71" }
 configparser = { version = "3.0.2", features = ["indexmap"] }
 queues = { version = "1.0.2" }
-assert_fs = { version = "1.0.13" }
 
 [dev-dependencies]
 assert_cmd = { version = "2.0.11" }
 predicates = { version = "3.0.3" }
+assert_fs = { version = "1.0.13" }
