+++ tests/cases/android/output1/values-ar/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<resources>
+  <string name="general">عام</string>
+  <string name="notifications">الإشعارات</string>
+</resources>
