+++ tests/cases/android/output1/values-fi/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<resources>
+  <string name="general">Yleistä</string>
+  <string name="notifications">Ilmoitukset</string>
+</resources>
