+++ tests/cases/android/output1/values-fr/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<resources>
+  <string name="general">Général</string>
+  <string name="notifications">Notifications</string>
+</resources>
